waifu/pyproject.toml
[tool.poetry]
name = "dicwaifu"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
asyncio = "^3.4.3"
transformers = "^4.45.2"
pillow = "^11.0.0"
aiohttp = "^3.10.10"
diffusers = "^0.31.0"
characterai = "^1.0.1"
openai = "^1.54.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
